backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
websockets>=11.0.0
requests>=2.31.0
python-reapy>=0.10.0
numpy>=1.24.0